pyproject.toml
[project]
name = "resume-bot"
version = "0.1.0"
description = "Resume Bot with FastAPI and LangGraph"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
    "langchain>=0.3.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.4.0",
    "langsmith>=0.3.0",
    "openai>=1.50.0",
    "requests>=2.32.0",
    "streamlit>=1.45.0",
    "ipykernel>=6.29.5",
]
